requirements.txt
liboqs-python>=0.7.2
pycryptodome>=3.18.0
cryptography>=41.0.0
pytest>=7.4.0
memory-profiler>=0.61.0
matplotlib>=3.7.0